ROX-6716: Allow wildcard certs to be used with non-helm-managed clusters (stackrox/rox#7732)

+++ sensor/kubernetes/upgrade/command_handler_impl.go
@@ -133,7 +133,7 @@ func (h *commandHandler) ProcessMessage(msg *central.MsgToSensor) error {
 		return nil
 	}
 
-	if h.configHandler.GetHelmManagedConfig() != nil {
+	if h.configHandler.GetHelmManagedConfig() != nil && !h.configHandler.GetHelmManagedConfig().GetNotHelmManaged() {
 		upgradesNotSupportedErr := errors.New("Cluster is Helm-managed and does not support auto-upgrades")
 		go h.rejectUpgradeRequest(trigger, upgradesNotSupportedErr)
 		go h.deleteUpgraderDeployments()
